requirements.txt
cryptography==3.3.1
Django==3.1.5
django-cors-headers==3.6.0
django-email-verification==0.0.7
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
google-auth==1.24.0
gunicorn==20.0.4
packaging==20.8
pandas==1.2.0
Pillow==7.1.2
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3

simplejwt==2.0.1

whitenoise==5.3.0
